Update to Mojolicious 7.30

+++ lib/Mojo/Server/Hypnotoad.pm
@@ -3,41 +3,58 @@ use Mojo::Base -base;
 
 # "Bender: I was God once.
 #  God: Yes, I saw. You were doing well, until everyone died."
-use Cwd 'abs_path';
-use File::Basename 'dirname';
-use File::Spec::Functions 'catfile';
+use Config;
+use Mojo::File 'path';
 use Mojo::Server::Prefork;
 use Mojo::Util 'steady_time';
-use POSIX 'setsid';
 use Scalar::Util 'weaken';
 
+has prefork => sub { Mojo::Server::Prefork->new(listen => ['http://*:8080']) };
+has upgrade_timeout => 60;
+
+sub configure {
+  my ($self, $name) = @_;
+
+  # Hypnotoad settings
+  my $prefork = $self->prefork;
+  my $c = $prefork->app->config($name) || {};
+  $self->upgrade_timeout($c->{upgrade_timeout}) if $c->{upgrade_timeout};
+
+  # Pre-fork settings
+  $prefork->reverse_proxy($c->{proxy})   if defined $c->{proxy};
+  $prefork->max_clients($c->{clients})   if $c->{clients};
+  $prefork->max_requests($c->{requests}) if $c->{requests};
+  defined $c->{$_} and $prefork->$_($c->{$_})
+    for qw(accepts backlog graceful_timeout heartbeat_interval),
+    qw(heartbeat_timeout inactivity_timeout listen pid_file workers);
+}
+
 sub run {
-  my ($self, $path) = @_;
+  my ($self, $app) = @_;
 
-  # No Windows support
-  _exit('Hypnotoad not available for Windows.') if $^O eq 'MSWin32';
+  # No fork emulation support
+  _exit('Hypnotoad does not support fork emulation.') if $Config{d_pseudofork};
 
-  # Remember application for later
-  $ENV{HYPNOTOAD_APP} ||= abs_path $path;
+  # Remember executable and application for later
+  $ENV{HYPNOTOAD_EXE} ||= $0;
+  $0 = $ENV{HYPNOTOAD_APP} ||= path($app)->to_abs->to_string;
 
   # This is a production server
   $ENV{MOJO_MODE} ||= 'production';
 
-  # Remember executable for later
-  $ENV{HYPNOTOAD_EXE} ||= $0;
-  $0 = $ENV{HYPNOTOAD_APP};
-
-  # Clean start
-  die "Can't exec: $!" if !$ENV{HYPNOTOAD_REV}++ && !exec $ENV{HYPNOTOAD_EXE};
+  # Clean start (to make sure everything works)
+  die "Can't exec: $!"
+    if !$ENV{HYPNOTOAD_REV}++ && !exec $^X, $ENV{HYPNOTOAD_EXE};
 
   # Preload application and configure server
-  my $prefork = $self->{prefork} = Mojo::Server::Prefork->new;
-  my $app = $prefork->load_app($ENV{HYPNOTOAD_APP});
-  $self->_config($app);
+  my $prefork = $self->prefork->cleanup(0);
+  $prefork->load_app($app)->config->{hypnotoad}{pid_file}
+    //= path($ENV{HYPNOTOAD_APP})->sibling('hypnotoad.pid')->to_string;
+  $self->configure('hypnotoad');
   weaken $self;
   $prefork->on(wait   => sub { $self->_manage });
-  $prefork->on(reap   => sub { $self->_reap(pop) });
-  $prefork->on(finish => sub { $self->{finished} = 1 });
+  $prefork->on(reap   => sub { $self->_cleanup(pop) });
+  $prefork->on(finish => sub { $self->_finish });
 
   # Testing
   _exit('Everything looks good!') if $ENV{HYPNOTOAD_TEST};
@@ -49,52 +66,40 @@ sub run {
   $self->_hot_deploy unless $ENV{HYPNOTOAD_PID};
 
   # Daemonize as early as possible (but not for restarts)
-  if (!$ENV{HYPNOTOAD_FOREGROUND} && $ENV{HYPNOTOAD_REV} < 3) {
-
-    # Fork and kill parent
-    die "Can't fork: $!" unless defined(my $pid = fork);
-    exit 0 if $pid;
-    setsid or die "Can't start a new session: $!";
-
-    # Close filehandles
-    open STDIN,  '</dev/null';
-    open STDOUT, '>/dev/null';
-    open STDERR, '>&STDOUT';
-  }
+  $prefork->start;
+  $prefork->daemonize if !$ENV{HYPNOTOAD_FOREGROUND} && $ENV{HYPNOTOAD_REV} < 3;
 
   # Start accepting connections
   local $SIG{USR2} = sub { $self->{upgrade} ||= steady_time };
-  $prefork->run;
+  $prefork->cleanup(1)->run;
 }
 
-sub _config {
-  my ($self, $app) = @_;
+sub _cleanup {
+  my ($self, $pid) = @_;
 
-  # Hypnotoad settings
-  my $c = $app->config('hypnotoad') || {};
-  $self->{upgrade_timeout} = $c->{upgrade_timeout} || 60;
-
-  # Prefork settings
-  $ENV{MOJO_REVERSE_PROXY} = $c->{proxy} if defined $c->{proxy};
-  my $prefork = $self->{prefork}->listen($c->{listen} || ['http://*:8080']);
-  my $file = catfile dirname($ENV{HYPNOTOAD_APP}), 'hypnotoad.pid';
-  $prefork->pid_file($c->{pid_file} || $file);
-  $prefork->max_clients($c->{clients}) if $c->{clients};
-  $prefork->max_requests($c->{keep_alive_requests})
-    if $c->{keep_alive_requests};
-  defined $c->{$_} and $prefork->$_($c->{$_})
-    for qw(accept_interval accepts backlog graceful_timeout group),
-    qw(heartbeat_interval heartbeat_timeout inactivity_timeout lock_file),
-    qw(lock_timeout multi_accept user workers);
+  # Clean up failed upgrade
+  return unless ($self->{new} || '') eq $pid;
+  $self->prefork->app->log->error('Zero downtime software upgrade failed');
+  delete @$self{qw(new upgrade)};
 }
 
 sub _exit { say shift and exit 0 }
 
-sub _hot_deploy {
+sub _finish {
   my $self = shift;
 
+  $self->{finish} = 1;
+  return unless my $new = $self->{new};
+
+  my $prefork = $self->prefork->cleanup(0);
+  unlink $prefork->pid_file;
+  $prefork->ensure_pid_file($new);
+}
+
+sub _hot_deploy {
+
   # Make sure server is running
-  return unless my $pid = $self->{prefork}->check_pid;
+  return unless my $pid = shift->prefork->check_pid;
 
   # Start hot deployment
   kill 'USR2', $pid;
@@ -104,10 +109,12 @@ sub _hot_deploy {
 sub _manage {
   my $self = shift;
 
-  # Upgraded
-  my $log = $self->{prefork}->app->log;
+  # Upgraded (wait for all workers to send a heartbeat)
+  my $prefork = $self->prefork;
+  my $log     = $prefork->app->log;
   if ($ENV{HYPNOTOAD_PID} && $ENV{HYPNOTOAD_PID} ne $$) {
-    $log->info("Upgrade successful, stopping $ENV{HYPNOTOAD_PID}.");
+    return unless $prefork->healthy == $prefork->workers;
+    $log->info("Upgrade successful, stopping $ENV{HYPNOTOAD_PID}");
     kill 'QUIT', $ENV{HYPNOTOAD_PID};
   }
   $ENV{HYPNOTOAD_PID} = $$ unless ($ENV{HYPNOTOAD_PID} // '') eq $$;
@@ -117,103 +124,94 @@ sub _manage {
 
     # Fresh start
     unless ($self->{new}) {
-      $log->info('Starting zero downtime software upgrade.');
+      $log->info('Starting zero downtime software upgrade');
       die "Can't fork: $!" unless defined(my $pid = $self->{new} = fork);
-      exec($ENV{HYPNOTOAD_EXE}) or die("Can't exec: $!") unless $pid;
+      exec $^X, $ENV{HYPNOTOAD_EXE} or die "Can't exec: $!" unless $pid;
     }
 
     # Timeout
     kill 'KILL', $self->{new}
-      if $self->{upgrade} + $self->{upgrade_timeout} <= steady_time;
+      if $self->{upgrade} + $self->upgrade_timeout <= steady_time;
   }
-}
-
-sub _reap {
-  my ($self, $pid) = @_;
-
-  # Clean up failed upgrade
-  return unless ($self->{new} || '') eq $pid;
-  $self->{prefork}->app->log->info('Zero downtime software upgrade failed.');
-  delete $self->{$_} for qw(new upgrade);
 }
 
 sub _stop {
   _exit('Hypnotoad server not running.')
-    unless my $pid = shift->{prefork}->check_pid;
+    unless my $pid = shift->prefork->check_pid;
   kill 'QUIT', $pid;
   _exit("Stopping Hypnotoad server $pid gracefully.");
 }
 
 1;
 
+=encoding utf8
+
 =head1 NAME
 
-Mojo::Server::Hypnotoad - ALL GLORY TO THE HYPNOTOAD!
+Mojo::Server::Hypnotoad - A production web serv...ALL GLORY TO THE HYPNOTOAD!
 
 =head1 SYNOPSIS
 
   use Mojo::Server::Hypnotoad;
 
-  my $toad = Mojo::Server::Hypnotoad->new;
-  $toad->run('/home/sri/myapp.pl');
+  my $hypnotoad = Mojo::Server::Hypnotoad->new;
+  $hypnotoad->run('/home/sri/myapp.pl');
 
 =head1 DESCRIPTION
 
-L<Mojo::Server::Hypnotoad> is a full featured, UNIX optimized, preforking
+L<Mojo::Server::Hypnotoad> is a full featured, UNIX optimized, pre-forking
 non-blocking I/O HTTP and WebSocket server, built around the very well tested
-and reliable L<Mojo::Server::Prefork>, with IPv6, TLS, Comet (long polling),
-keep-alive, connection pooling, timeout, cookie, multipart, multiple event
-loop and hot deployment support that just works. Note that the server uses
-signals for process management, so you should avoid modifying signal handlers
-in your applications.
+and reliable L<Mojo::Server::Prefork>, with IPv6, TLS, SNI, UNIX domain socket,
+Comet (long polling), keep-alive, multiple event loop and hot deployment support
+that just works. Note that the server uses signals for process management, so
+you should avoid modifying signal handlers in your applications.
 
-To start applications with it you can use the L<hypnotoad> script.
+To start applications with it you can use the L<hypnotoad> script, which
+listens on port C<8080>, automatically daemonizes the server process and
+defaults to C<production> mode for L<Mojolicious> and L<Mojolicious::Lite>
+applications.
 
-  $ hypnotoad myapp.pl
-  Server available at http://127.0.0.1:8080.
+  $ hypnotoad ./myapp.pl
 
 You can run the same command again for automatic hot deployment.
 
-  $ hypnotoad myapp.pl
+  $ hypnotoad ./myapp.pl
   Starting hot deployment for Hypnotoad server 31841.
 
-For L<Mojolicious> and L<Mojolicious::Lite> applications it will default to
-C<production> mode.
+This second invocation will load the application again, detect the process id
+file with it, and send a L</"USR2"> signal to the already running server.
 
-For better scalability (epoll, kqueue) and to provide IPv6 as well as TLS
-support, the optional modules L<EV> (4.0+), L<IO::Socket::IP> (0.16+) and
-L<IO::Socket::SSL> (1.75+) will be used automatically by L<Mojo::IOLoop> if
-they are installed. Individual features can also be disabled with the
-MOJO_NO_IPV6 and MOJO_NO_TLS environment variables.
+For better scalability (epoll, kqueue) and to provide non-blocking name
+resolution, SOCKS5 as well as TLS support, the optional modules L<EV> (4.0+),
+L<Net::DNS::Native> (0.15+), L<IO::Socket::Socks> (0.64+) and
+L<IO::Socket::SSL> (1.94+) will be used automatically if possible. Individual
+features can also be disabled with the C<MOJO_NO_NNR>, C<MOJO_NO_SOCKS> and
+C<MOJO_NO_TLS> environment variables.
 
-See L<Mojolicious::Guides::Cookbook> for more.
+See L<Mojolicious::Guides::Cookbook/"DEPLOYMENT"> for more.
 
-=head1 SIGNALS
+=head1 MANAGER SIGNALS
 
-L<Mojo::Server::Hypnotoad> can be controlled at runtime with the following
-signals.
+The L<Mojo::Server::Hypnotoad> manager process can be controlled at runtime
+with the following signals.
 
-=head2 Manager
+=head2 INT, TERM
 
-=over 2
+Shut down server immediately.
 
-=item INT, TERM
+=head2 QUIT
 
-Shutdown server immediately.
+Shut down server gracefully.
 
-=item QUIT
-
-Shutdown server gracefully.
-
-=item TTIN
+=head2 TTIN
 
 Increase worker pool by one.
 
-=item TTOU
+=head2 TTOU
 
 Decrease worker pool by one.
 
-=item USR2
+=head2 USR2
 
 Attempt zero downtime software upgrade (hot deployment) without losing any
 incoming connections.
@@ -229,130 +227,93 @@ incoming connections.
      |- Worker [3]
      +- Worker [4]
 
-The new manager will automatically send a C<QUIT> signal to the old manager
+The new manager will automatically send a L</"QUIT"> signal to the old manager
 and take over serving requests after starting up successfully.
 
-=back
+=head1 WORKER SIGNALS
 
-=head2 Worker
+L<Mojo::Server::Hypnotoad> worker processes can be controlled at runtime with
+the following signals.
 
-=over 2
-
-=item INT, TERM
-
-Stop worker immediately.
-
-=item QUIT
+=head2 QUIT
 
 Stop worker gracefully.
-
-=back
 
 =head1 SETTINGS
 
 L<Mojo::Server::Hypnotoad> can be configured with the following settings, see
 L<Mojolicious::Guides::Cookbook/"Hypnotoad"> for examples.
 
-=head2 accept_interval
-
-  accept_interval => 0.5
-
-Interval in seconds for trying to reacquire the accept mutex, defaults to
-C<0.025>. Note that changing this value can affect performance and idle CPU
-usage.
-
 =head2 accepts
 
   accepts => 100
 
-Maximum number of connections a worker is allowed to accept before stopping
-gracefully, defaults to C<1000>. Setting the value to C<0> will allow workers
-to accept new connections indefinitely. Note that up to half of this value can
-be subtracted randomly to improve load balancing, and that worker processes
-will stop sending heartbeat messages once the limit has been reached.
+Maximum number of connections a worker is allowed to accept, before stopping
+gracefully and then getting replaced with a newly started worker, defaults to
+the value of L<Mojo::Server::Prefork/"accepts">. Setting the value to C<0> will
+allow workers to accept new connections indefinitely. Note that up to half of
+this value can be subtracted randomly to improve load balancing, and to make
+sure that not all workers restart at the same time.
 
 =head2 backlog
 
   backlog => 128
 
-Listen backlog size, defaults to C<SOMAXCONN>.
+Listen backlog size, defaults to the value of
+L<Mojo::Server::Daemon/"backlog">.
 
 =head2 clients
 
   clients => 100
 
-Maximum number of parallel client connections per worker process, defaults to
-C<1000>. Note that depending on how much your application may block, you might
-want to decrease this value and increase C<workers> instead for better
-performance.
+Maximum number of accepted connections each worker process is allowed to handle
+concurrently, before stopping to accept new incoming connections, defaults to
+the value of L<Mojo::IOLoop/"max_connections">. Note that high concurrency works
+best with applications that perform mostly non-blocking operations, to optimize
+for blocking operations you can decrease this value and increase L</"workers">
+instead for better performance.
 
 =head2 graceful_timeout
 
   graceful_timeout => 15
 
 Maximum amount of time in seconds stopping a worker gracefully may take before
-being forced, defaults to C<20>.
-
-=head2 group
-
-  group => 'staff'
-
-Group name for worker processes.
+being forced, defaults to the value of
+L<Mojo::Server::Prefork/"graceful_timeout">. Note that this value should usually
+be a little larger than the maximum amount of time you expect any one request to
+take.
 
 =head2 heartbeat_interval
 
   heartbeat_interval => 3
 
-Heartbeat interval in seconds, defaults to C<5>.
+Heartbeat interval in seconds, defaults to the value of
+L<Mojo::Server::Prefork/"heartbeat_interval">.
 
 =head2 heartbeat_timeout
 
   heartbeat_timeout => 2
 
 Maximum amount of time in seconds before a worker without a heartbeat will be
-stopped gracefully, defaults to C<20>.
+stopped gracefully, defaults to the value of
+L<Mojo::Server::Prefork/"heartbeat_timeout">. Note that this value should
+usually be a little larger than the maximum amount of time you expect any one
+operation to block the event loop.
 
 =head2 inactivity_timeout
 
   inactivity_timeout => 10
 
 Maximum amount of time in seconds a connection can be inactive before getting
-closed, defaults to C<15>. Setting the value to C<0> will allow connections to
-be inactive indefinitely.
-
-=head2 keep_alive_requests
-
-  keep_alive_requests => 50
-
-Number of keep-alive requests per connection, defaults to C<25>.
+closed, defaults to the value of L<Mojo::Server::Daemon/"inactivity_timeout">.
+Setting the value to C<0> will allow connections to be inactive indefinitely.
 
 =head2 listen
 
   listen => ['http://*:80']
 
-List of one or more locations to listen on, defaults to C<http://*:8080>. See
-also L<Mojo::Server::Daemon/"listen"> for more examples.
-
-=head2 lock_file
-
-  lock_file => '/tmp/hypnotoad.lock'
-
-Full path of accept mutex lock file prefix, to which the process id will be
-appended, defaults to a random temporary path.
-
-=head2 lock_timeout
-
-  lock_timeout => 0.5
-
-Maximum amount of time in seconds a worker may block when waiting for the
-accept mutex, defaults to C<1>. Note that changing this value can affect
-performance and idle CPU usage.
-
-=head2 multi_accept
-
-  multi_accept => 100
-
-Number of connections to accept at once, defaults to C<50>.
+Array reference with one or more locations to listen on, defaults to
+C<http://*:8080>. See also L<Mojo::Server::Daemon/"listen"> for more examples.
 
 =head2 pid_file
 
@@ -367,8 +328,15 @@ the server has been stopped.
   proxy => 1
 
 Activate reverse proxy support, which allows for the C<X-Forwarded-For> and
-C<X-Forwarded-HTTPS> headers to be picked up automatically, defaults to the
-value of the MOJO_REVERSE_PROXY environment variable.
+C<X-Forwarded-Proto> headers to be picked up automatically, defaults to the
+value of L<Mojo::Server/"reverse_proxy">.
+
+=head2 requests
+
+  requests => 50
+
+Number of keep-alive requests per connection, defaults to the value of
+L<Mojo::Server::Daemon/"max_requests">.
 
 =head2 upgrade_timeout
 
@@ -377,32 +345,56 @@ value of the MOJO_REVERSE_PROXY environment variable.
 Maximum amount of time in seconds a zero downtime software upgrade may take
 before getting canceled, defaults to C<60>.
 
-=head2 user
-
-  user => 'sri'
-
-Username for worker processes.
-
 =head2 workers
 
   workers => 10
 
-Number of worker processes, defaults to C<4>. A good rule of thumb is two
-worker processes per CPU core.
+Number of worker processes, defaults to the value of
+L<Mojo::Server::Prefork/"workers">. A good rule of thumb is two worker
+processes per CPU core for applications that perform mostly non-blocking
+operations, blocking operations often require more and benefit from decreasing
+concurrency with L</"clients"> (often as low as C<1>). Note that during zero
+downtime software upgrades there will be twice as many workers active for a
+short amount of time.
+
+=head1 ATTRIBUTES
+
+L<Mojo::Server::Hypnotoad> implements the following attributes.
+
+=head2 prefork
+
+  my $prefork = $hypnotoad->prefork;
+  $hypnotoad  = $hypnotoad->prefork(Mojo::Server::Prefork->new);
+
+L<Mojo::Server::Prefork> object this server manages.
+
+=head2 upgrade_timeout
+
+  my $timeout = $hypnotoad->upgrade_timeout;
+  $hypnotoad  = $hypnotoad->upgrade_timeout(15);
+
+Maximum amount of time in seconds a zero downtime software upgrade may take
+before getting canceled, defaults to C<60>.
 
 =head1 METHODS
 
 L<Mojo::Server::Hypnotoad> inherits all methods from L<Mojo::Base> and
 implements the following new ones.
 
+=head2 configure
+
+  $hypnotoad->configure('hypnotoad');
+
+Configure server from application settings.
+
 =head2 run
 
-  $toad->run('script/myapp');
+  $hypnotoad->run('script/my_app');
 
-Run server for application.
+Run server for application and wait for L</"MANAGER SIGNALS">.
 
 =head1 SEE ALSO
 
-L<Mojolicious>, L<Mojolicious::Guides>, L<http://mojolicio.us>.
+L<Mojolicious>, L<Mojolicious::Guides>, L<http://mojolicious.org>.
 
 =cut
